requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
statsmodels
pmdarima
matplotlib
plotly
joblib
PyYAML
fastapi
uvicorn
mlflow
